[resource_groups] move verifier code
