+++ aptos-move/aptos-vm/src/verifier/mod.rs
@@ -1,5 +1,6 @@
 // Copyright (c) Aptos
 // SPDX-License-Identifier: Apache-2.0
 pub(crate) mod module_init;
+pub(crate) mod resource_groups;
 pub mod transaction_arg_validation;
 pub(crate) mod view_function;
